James's suggestions
