+++ web/.gitignore
@@ -5,3 +5,4 @@ out
 tsconfig.tsbuildinfo
 .env*
 .vercel
+testing
