+++ web/hooks/use-saved-contract-metrics.ts
@@ -7,7 +7,9 @@ import { safeLocalStorage } from 'web/lib/util/local'
 
 export const useSavedContractMetrics = (contract: Contract) => {
   const user = useUser()
-  const contractMetrics = useUserContractMetrics(user?.id, contract.id)
+  const contractMetrics = contract
+    ? useUserContractMetrics(user?.id, contract.id)
+    : null
   const [savedMetrics, setSavedMetrics] = usePersistentState<
     ContractMetrics | undefined
   >(undefined, {
